Twitter-Clone
This project is a Twitter-Clone built using a full-stack approach, leveraging React for the frontend and other backend technologies for seamless functionality and scalability. The clone aims to replicate key features of Twitter, providing users with the ability to post tweets, follow other users, engage with content, and enjoy a streamlined social media experience.